fix broker test
thought that the broker timeout was causing the flaky tests. Turns
out that the timeout would cause a timeout exception, not a cancellation
exception, which is what I was seeing. It looks like when you cancel
the broker server, you might occasionally get something that's not a
grpc.StatusException. I'm not sure what it is, because the test isn't
flaky for me locally.